VERY IMPORTANT! CB_Activity sheet needs to be shared Anyone With Link
If you don't, the web app runs a very high risk of eventually stop working, as the API call for the query will require credentials that can't be provided. When source sheet is shared Anyone with link, the API call isn't connected to a specific user, and circumvents this requirement.